async: more featured async consumer
The previous consume! function makes it really hard to perform in-flight operations on a consumer since consumption cannot happen on a separate thread.

The async facade now accounts for that and completely does away with exposing the consumer driver.
Instead it yields a vector [records ctl] of two channels, one for consuming output records and control messages, one for interacting with the consumer driver.

The call signature remains similar to the previous one, additional options may be present in the configuration to control buffer sizes and polling timeout.

The control messages understood are: pause, resume, subscribe, unsubscribe, commit, partitions-for, and stop. A callback message can also be provided to execute code on a stopped consumer driver.

This will deprecate consume!

The PR also does a drive-by downgrade of the underlying library from 0.10.0.1 to 0.9.0.1 since there is no descending compatibility and most people are unlikely to be on 0.10.x yet.