Fix async consumer ConcurrentModificationException (#39)
* Prepare for 0.1.23

* Async consumer: fix ConcurrentModificationException bug

When creating the consumer given some topic to subscribe to, the
subscription operation was done by calling the client driver directly,
which has the adverse effect of pinning the underlying kafka consumer
to the calling thread, thus making the poller thread throw a
ConcurrentModificationException. Instead of calling the driver, send
a subscribe operation.