Initial ideas for an `AdminClient` wrapper (#30)
* Initial ideas for an `AdminClient` wrapper

This is an initial PR to get some early feedback on the approach. I've
tried to adapt to the existing style and code in the `client` namespace.

The `AdminClient` is fully async and returning instances of
`KafkaFuture`s. This PR contains a wrapper around this that allows
dereferencing the result (tested) and use `core.async` channels
(untested so far).

Refs #29

* Use extend-type for the AdminClient directly

Instead of creating a wrapper object around the `AdminClient` we can
simply extend it with the `AdminClientDriver` protocol.

* Add topic creation

Rough edges for topic creation.